pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance-agent-baseline"
version = "1.0.0"
description = "Finance Agent Benchmark 2.0 - Purple Agent (Baseline)"
authors = [
    {name = "Finance Agent Benchmark Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.5",
    "pydantic>=2.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.35.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
